// src/generated/trading212.ts
export interface paths {
// ...
    "/api/v0/equity/history/dividends": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Get paid out dividends
         * @description **Rate limit:** 6 req / 1m0s
         */
        get: operations["dividends"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/api/v0/equity/history/exports": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * List generated reports
         * @description Retrieves a list of all requested CSV reports and their current status.
         *
         *
         *     **Asynchronous Workflow:**
         *
         *     1. Call `POST /history/exports` to request a report. You will receive a
         *     `reportId`.
         *
         *     2. Periodically call this endpoint (`GET /history/exports`) to check the
         *     `status` of the report corresponding to your `reportId`.
         *
         *     3. Once the status is `Finished`, the `downloadLink` field will contain
         *     a URL to download the CSV file.
         *
         *     **Rate limit:** 1 req / 1m0s
         */
        get: operations["getReports"];
        put?: never;
        /**
         * Request a CSV report
         * @description Initiates the generation of a CSV report containing historical account
         *     data. This is an asynchronous operation. The response will include a
         *     `reportId` which you can use to track the status of the generation
         *     process using the `GET /history/exports` endpoint.
         *
         *     **Rate limit:** 1 req / 30s
         */
        post: operations["requestReport"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/api/v0/equity/history/orders": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Get historical orders data
         * @description **Rate limit:** 6 req / 1m0s
         */
        get: operations["orders_1"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/api/v0/equity/history/transactions": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Get transactions
         * @description Fetch superficial information about movements to and from your account
         *
         *     **Rate limit:** 6 req / 1m0s
         */
        get: operations["transactions"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
}
// ...
export interface operations {
    getAccountSummary: {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
// ...
    };
// ...
}